requirements.txt
openai-whisper>=20240101
numpy>=1.20.0
scipy>=1.7.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
pillow>=10.0.0
sounddevice>=0.4.5
pygame>=2.5.0
soundfile>=0.12.0
pywin32>=305; platform_system == "Windows"
winshell>=0.6; platform_system == "Windows"
